spotless + add mapping verification test

+++ src/main/java/org/opensearch/knn/index/codec/params/KNNVectorsFormatParams.java
@@ -7,7 +7,6 @@ package org.opensearch.knn.index.codec.params;
 
 import lombok.Getter;
 import org.opensearch.knn.common.KNNConstants;
-import org.opensearch.knn.index.KNNSettings;
 import org.opensearch.knn.index.SpaceType;
 
 import java.util.Map;
@@ -28,8 +27,8 @@ public class KNNVectorsFormatParams {
             params,
             defaultMaxConnections,
             defaultBeamWidth,
-                KNNConstants.DEFAULT_ALPHA_VALUE.floatValue(),
-                KNNConstants.DEFAULT_NEIGHBOR_OVERFLOW_VALUE.floatValue(),
+            KNNConstants.DEFAULT_ALPHA_VALUE.floatValue(),
+            KNNConstants.DEFAULT_NEIGHBOR_OVERFLOW_VALUE.floatValue(),
             SpaceType.UNDEFINED
         );
     }
@@ -71,7 +70,7 @@ public class KNNVectorsFormatParams {
 
     private void initAlpha(final Map<String, Object> params, float defaultAlpha) {
         if (params != null && params.containsKey(KNNConstants.METHOD_PARAMETER_ALPHA)) {
-            this.alpha = (int) params.get(KNNConstants.METHOD_PARAMETER_ALPHA);
+            this.alpha = ((Double) params.get(KNNConstants.METHOD_PARAMETER_ALPHA)).floatValue();
             return;
         }
         this.alpha = defaultAlpha;
@@ -79,7 +78,7 @@ public class KNNVectorsFormatParams {
 
     private void initNeighborOverflow(final Map<String, Object> params, float defaultNeighborOverflow) {
         if (params != null && params.containsKey(KNNConstants.METHOD_PARAMETER_NEIGHBOR_OVERFLOW)) {
-            this.neighborOverflow = (int) params.get(KNNConstants.METHOD_PARAMETER_NEIGHBOR_OVERFLOW);
+            this.neighborOverflow = ((Double) params.get(KNNConstants.METHOD_PARAMETER_NEIGHBOR_OVERFLOW)).floatValue();
             return;
         }
         this.neighborOverflow = defaultNeighborOverflow;
